Moved app into features
